requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
umap-learn>=0.5.3
plotly>=5.0.0
opencv-python>=4.6.0
matplotlib>=3.5.0
joblib>=1.2.0
scipy>=1.9.0 